examples/tutorial/part0/requirements.txt
aiosqlite==0.19.0
black==23.12.1
cannula==0.10.0
click==8.1.7
fastapi[all]==0.108.0
pydantic_settings==2.1.0
sqlalchemy==2.0.25
uvicorn==0.25.0

# Testing
pytest<8
pytest-asyncio
pytest-cov
pytest-httpx
pytest-mock
